Refactor to improve readability and modularity
